fix: move report customizations to separate file in sample project
